+++ src/shared/components/palette/palette.service.ts
@@ -44,6 +44,7 @@ import { ReadOrganisationFieldComponent } from './organisation/read-organisation
 import { ReadDynamicRadioListFieldComponent } from './dynamic-radio-list/read-dynamic-radio-list-field.component';
 import { WriteDynamicRadioListFieldComponent } from './dynamic-radio-list/write-dynamic-radio-list-field.component';
 import { WaysToPayFieldComponent } from './waystopay/waystopay-field.component';
+import { WriteDynamicMultiSelectListFieldComponent, ReadDynamicMultiSelectListFieldComponent } from './dynamic-multi-select-list';
 
 @Injectable()
 export class PaletteService {
@@ -74,6 +75,8 @@ export class PaletteService {
         return write ? WriteFixedListFieldComponent : ReadFixedListFieldComponent;
       case 'DynamicRadioList':
         return write ? WriteDynamicRadioListFieldComponent : ReadDynamicRadioListFieldComponent;
+      case 'DynamicMultiSelectList':
+        return write ? WriteDynamicMultiSelectListFieldComponent : ReadDynamicMultiSelectListFieldComponent;
       case 'FixedRadioList':
         return write ? WriteFixedRadioListFieldComponent : ReadFixedRadioListFieldComponent;
       case 'Complex':