+++ RELEASE-NOTES.md
@@ -1,5 +1,7 @@
 ## RELEASE NOTES
 
+### Version 4.18.16-EUI-2831-dynamic-multiselect
+**EUI-2831** Dynamic multi-select list
 
 ### Version 4.12.1-EUI-5225-organisation-search
 **EUI-5225** Bug fix the organisation search functionality